move `PrefetchDirection` to types file; update imports
