+++ src/test/ChunkPrefetchIterator.test.ts
@@ -1,7 +1,8 @@
 import { expect } from "chai";
 
 import { TCZYX } from "../loaders/zarr_utils/types";
-import ChunkPrefetchIterator, { PrefetchDirection } from "../loaders/zarr_utils/ChunkPrefetchIterator";
+import ChunkPrefetchIterator from "../loaders/zarr_utils/ChunkPrefetchIterator";
+import { PrefetchDirection } from "../loaders/zarr_utils/types";
 
 const EXPECTED_3X3X3X3 = [
   [0, 0, 1, 1, 1], // T-
